Repeat prompts should not be interruptible
In cases where the first response to a prompt is not a match, it is likely that the issue was with background noise or excessive sensitivity. Such conditions with repeat prompts can lead to a bouncing behaviour. Removing the potential entirely for feedback of the prompt into input can help here, but even in cases where it doesn't, the absence of very short repeat prompts (of circa 1 second) makes for a more graceful failure.